requirements.txt
datasets~=2.13.1
faiss-cpu~=1.7.4
gensim~=4.3.2
lightning~=2.0.6
matplotlib~=3.7.2
networkx~=3.1
nltk~=3.8.1
numpy~=1.25.0
openai~=0.27.8
pandas~=1.4.3
pykeen==1.10.1
python-dotenv~=1.0.0
scikit-learn~=1.3.0
sentence-transformers==2.2.2
spacy~=3.5.4
torch~=2.0.1
torch-geometric==2.3.1
transformers~=4.30.2
tsne-torch~=1.0.1
tqdm~=4.65.0
wordcloud==1.9.2